Refatoração da 017

+++ CEV/017.py
@@ -2,7 +2,7 @@ from math import hypot
 
 print('O valor da hipotenusa é: {:.2f}'.format(
     hypot(
-        (int(input('Digite o valor do cateto oposto: '))),
-        (int(input('Digite o valor do cateto adjacente: ')))
+        (float(input('Digite o valor do cateto oposto: '))),
+        (float(input('Digite o valor do cateto adjacente: ')))
         )
 ))
